Run simple ring-app on uWSGI server.
Just was curious how to run Clojure/Ring application on uWSGI, it was
written http://uwsgi-docs.readthedocs.org/en/latest/Ring.html that it is
possible to do, but after all some tweaks were required (see README.md
for the details) to make it work.

The app just display the factorial of n (localhost:9090/?n=8) or
factorial of 10 by default.