Tests are passing now.
Did mocking with 'with-redefs', it works, but is this the clojure way to
do mock testing?
Of course, mock testing assumes white box opposed to black box, because
we mock actual implementation details and if we change the
implementation which still can produce the correct result, but test will
fail, because we use different functions to complete the job.

And we need to update the test in that case. Of, course much easier is
to test "pure" functions, without side effects.

Still something investigate on clojure testing layer.