73, Analyze a Tic-Tac-Toe board.
4clojure puts some restrictions on the implementations, apparently
everything must be done in one function, and no nested #() anon
functions are allowed, so need to use (fn []) instead.

Also my golf score is not perfect. Is it an issue? Readability counts.
I'm for readability, simplicity and ease of understanding, rather than make it as smaller as
possible. Although worth to investigate other solutions.