One more step closer - added all-sub-squares fn.
This fn will generate all possible sub squares of the given dimension,
and then latin-square? will be used to check if the generated sub square
is the latin square or not. Also check for 'X symbol must be done, to
skip illegal sub squares, because we are processing on the normalized V
(see normalize fn).