All parts are almost ready to start. 3, 2, 1, ... go :)
Added probably the last piece of puzzle into the place
next-shifted-square, right now need to use it to solve the problem.
Right now there is a huge mix of fn-s and their tests, need to split it
between 2 files probably, but then it will lack of input and output fn
expects and produces.