Refactored in-sequence? and introduced all-same-suites?
To be consistent and have all the functions to work with cards as an
input, changed signature of in-sequence? and introduced
all-same-suites?, because it is used 2 times already.

Of, course, in all helper functions I need to call ranks and/or suites,
unless there is a performance issue, I would prefer call them and have
more consistent and readable code.