Solved pr. 91: Graph Connectivity.
Although I don't like that I need to use atom #{} for that purpose. But
maybe sometime to do pure way is very difficult, maybe I can figure out
another solution with using just function constructs and not atom
(although atom is not bad I guess, it is after all part of the
language).